document the asynchronous API

+++ rodbus-ffi/src/asynchronous.rs
@@ -33,7 +33,7 @@ unsafe fn get_callback_session<'a>(
 
 unsafe fn data_callback_to_fn<T>(
     user_data: *mut c_void,
-    callback: Option<unsafe extern "C" fn(Result, *const T, usize, *mut c_void)>,
+    callback: Option<unsafe extern "C" fn(Result, *const T, u16, *mut c_void)>,
 ) -> impl Fn(std::result::Result<Vec<rodbus::types::Indexed<T>>, rodbus::error::Error>) -> ()
 where
     T: Copy,
@@ -48,7 +48,7 @@ where
                     cb(
                         Result::status(Status::Ok),
                         transformed.as_ptr(),
-                        transformed.len(),
+                        transformed.len() as u16,
                         user_data.value,
                     )
                 }
@@ -69,12 +69,21 @@ unsafe fn status_callback_to_fn<T>(
     }
 }
 
+/// @brief perform a non-blocking operation to read coils
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address for the operation
+/// @param count count of items for the operation
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn read_coils_cb(
     session: *mut Session,
     start: u16,
     count: u16,
-    callback: Option<unsafe extern "C" fn(Result, *const bool, usize, *mut c_void)>,
+    callback: Option<unsafe extern "C" fn(Result, *const bool, u16, *mut c_void)>,
     user_data: *mut c_void,
 ) {
     let (runtime, mut session) = get_callback_session(session);
@@ -85,12 +94,21 @@ pub unsafe extern "C" fn read_coils_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to read discrete inputs
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address for the operation
+/// @param count count of items for the operation
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn read_discrete_inputs_cb(
     session: *mut Session,
     start: u16,
     count: u16,
-    callback: Option<unsafe extern "C" fn(Result, *const bool, usize, *mut c_void)>,
+    callback: Option<unsafe extern "C" fn(Result, *const bool, u16, *mut c_void)>,
     user_data: *mut c_void,
 ) {
     let (runtime, mut session) = get_callback_session(session);
@@ -101,12 +119,21 @@ pub unsafe extern "C" fn read_discrete_inputs_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to read holding registers
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address for the operation
+/// @param count count of items for the operation
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn read_holding_registers_cb(
     session: *mut Session,
     start: u16,
     count: u16,
-    callback: Option<unsafe extern "C" fn(Result, *const u16, usize, *mut c_void)>,
+    callback: Option<unsafe extern "C" fn(Result, *const u16, u16, *mut c_void)>,
     user_data: *mut c_void,
 ) {
     let (runtime, mut session) = get_callback_session(session);
@@ -117,12 +144,21 @@ pub unsafe extern "C" fn read_holding_registers_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to read input registers
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address for the operation
+/// @param count count of items for the operation
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn read_input_registers_cb(
     session: *mut Session,
     start: u16,
     count: u16,
-    callback: Option<unsafe extern "C" fn(Result, *const u16, usize, *mut c_void)>,
+    callback: Option<unsafe extern "C" fn(Result, *const u16, u16, *mut c_void)>,
     user_data: *mut c_void,
 ) {
     let (runtime, mut session) = get_callback_session(session);
@@ -133,6 +169,15 @@ pub unsafe extern "C" fn read_input_registers_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to write a single coil
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param index address of the value
+/// @param value value to write
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn write_single_coil_cb(
     session: *mut Session,
@@ -149,6 +194,15 @@ pub unsafe extern "C" fn write_single_coil_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to write a single register
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param index address of the value
+/// @param value value to write
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data data providing context to the callback
+///
+/// @note This function is thread-safe
 #[no_mangle]
 pub unsafe extern "C" fn write_single_register_cb(
     session: *mut Session,
@@ -165,6 +219,18 @@ pub unsafe extern "C" fn write_single_register_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to write multiple coils
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address of the values
+/// @param values array of values to write
+/// @param count count of values to write
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data data providing context to the callback
+///
+/// @note This function is thread-safe
+/// @warning "values" must contain at least "count" items or the function
+/// will read past the end of the buffer
 #[no_mangle]
 pub unsafe extern "C" fn write_multiple_coils_cb(
     session: *mut Session,
@@ -182,6 +248,18 @@ pub unsafe extern "C" fn write_multiple_coils_cb(
     );
 }
 
+/// @brief perform a non-blocking operation to write multiple registers
+///
+/// @param session pointer to the #Session struct that provides the runtime, channel, etc
+/// @param start starting address of the values
+/// @param values array of values to write
+/// @param count count of values to write
+/// @param callback callback function to invoke when the operation completes
+/// @param user_data pointer to optional user data data providing context to the callback
+///
+/// @note This function is thread-safe
+/// @warning "values" must contain at least "count" items or the function
+/// will read past the end of the buffer
 #[no_mangle]
 pub unsafe extern "C" fn write_multiple_registers_cb(
     session: *mut Session,
